+++ tests/test_importing.py
@@ -5,6 +5,7 @@ def test_imports():
     """
     from tg_utils import admin
     from tg_utils import checks
+    from tg_utils import compressor_filters
     from tg_utils import email
     from tg_utils import files
     from tg_utils import hashmodels
@@ -13,4 +14,3 @@ def test_imports():
     from tg_utils import profiling
     from tg_utils import signals
     from tg_utils import uuid
-    from tg_utils import yuglify