+++ HISTORY.rst
@@ -2,6 +2,14 @@
 History
 =======
 
+0.2.0 (2016-01-27)
+------------------
+
+* Rename `tg_utils.yuglify` module to `tg_utils.compressor_filters` and
+  added new filters for using UglifyJS 2 and clean-css.
+* Added docs for `tg_utils.compressor_filters` module.
+
+
 0.1.1 (2016-01-11)
 ------------------